move files by catalogy and fix error by patH
