--- src/components/cardCatalogy/index.tsx
@@ -1,1 +1,0 @@
-export { CatalogyCard } from './ui';
