--- src/components/categoryList/index.tsx
@@ -1,1 +1,0 @@
-export { CategoryList } from './ui/CategoryList';
