--- src/components/filterCatalogy/index.tsx
@@ -1,1 +1,0 @@
-export { FilterCatalogySearch } from './ui/FilterCatalogy';
