--- src/components/pages/Catalogy/index.tsx
@@ -1,1 +1,0 @@
-export { CatalogyPage } from "./ui/CatalogyPage";
