--- src/components/pages/index.tsx
@@ -1,1 +1,0 @@
-export { CatalogyPage } from "./Catalogy";
